Turned accordion to be a shared component
